+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 1.0.3 (2016-12-22)
 
 -   Make repo in front end controller protected rather than private
+-   Fix user field default value and make defaulting to logged in user optional
 
 ## Version 1.0.2 (2016-12-21)